+++ src/app/features/products/components/product-list/product-list.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { Observable } from 'rxjs';
 import { CartService } from 'src/app/features/cart/services/cart.service';
 import { ProductModel } from '../../models/product';
 import { ProductsService } from '../../services/products.service';
@@ -10,7 +11,7 @@ import { ProductsService } from '../../services/products.service';
 })
 export class ProductListComponent implements OnInit {
 
-  products: ProductModel[];
+  products$: Observable<ProductModel[]>;
 
   constructor(
     private productsService: ProductsService,
@@ -19,13 +20,10 @@ export class ProductListComponent implements OnInit {
 
 
   ngOnInit(): void {
-    this.products = this.productsService.getProducts();
+    this.products$ = this.productsService.getProducts();
   }
 
-  onAddToCart(productId: number): void {
-    const product = this.products.find(x => x.id == productId);
-    if (product) {
-      this.cartService.addProduct(product);
-    }
+  onAddToCart(product: ProductModel): void {
+    this.cartService.addProduct(product);
   }
 }