+++ changelog.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2021-10-03
+### Removed
+### Changed
+- getProducts returns Observable
+- onAddToCart accepts ProductModel instead of productId
+### Added
+- Titlecase for cart & product items name
+
 ## [0.3.0] - 2021-09-27
 ### Removed
 ### Changed